+++ client/src/components/inventories/InventoryDelete.scss
@@ -1,0 +1,9 @@
+@import "../../partials/mixins.scss";
+@import "../../partials/partials.scss";
+
+.modal-custom {
+    background-color: green;
+    width: 10rem;
+    height: 50rem;
+    z-index: 3;
+}
